pyproject.toml
[project]
name = "financeflow-llm-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "amicus-veritatis",email = "otome@u.sogang.ac.kr"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langchain[openai] (>=0.3.25,<0.4.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)"
]
packages = [
    { include = "financeflow-llm-backend" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

